+++ src/Configuration.php
@@ -30,6 +30,8 @@ use Doctrine\Common\Collections\ArrayCollection;
 use TechDivision\Import\ConfigurationInterface;
 use TechDivision\Import\Configuration\DatabaseConfigurationInterface;
 use TechDivision\Import\Configuration\Jms\Configuration\Operation;
+use TechDivision\Import\Configuration\Jms\Configuration\ParamsTrait;
+use TechDivision\Import\Configuration\Jms\Configuration\CsvTrait;
 
 /**
  * A simple JMS based configuration implementation.
@@ -51,6 +53,20 @@ class Configuration implements ConfigurationInterface
      * @var string
      */
     const PID_FILENAME = 'importer.pid';
+
+    /**
+     * Trait that provides CSV configuration functionality.
+     *
+     * @var \TechDivision\Import\Configuration\Jms\Configuration\CsvTrait
+     */
+    use CsvTrait;
+
+    /**
+     * Trait that provides CSV configuration functionality.
+     *
+     * @var \TechDivision\Import\Configuration\Jms\Configuration\ParamsTrait
+     */
+    use ParamsTrait;
 
     /**
      * Mapping for boolean values passed on the console.
@@ -198,57 +214,6 @@ class Configuration implements ConfigurationInterface
      * @SerializedName("multiple-value-delimiter")
      */
     protected $multipleValueDelimiter = '|';
-
-    /**
-     * The subject's delimiter character for CSV files.
-     *
-     * @var string
-     * @Type("string")
-     */
-    protected $delimiter;
-
-    /**
-     * The subject's enclosure character for CSV files.
-     *
-     * @var string
-     * @Type("string")
-     */
-    protected $enclosure;
-
-    /**
-     * The subject's escape character for CSV files.
-     *
-     * @var string
-     * @Type("string")
-     */
-    protected $escape;
-
-    /**
-     * The subject's source charset for the CSV file.
-     *
-     * @var string
-     * @Type("string")
-     * @SerializedName("from-charset")
-     */
-    protected $fromCharset;
-
-    /**
-     * The subject's target charset for a CSV file.
-     *
-     * @var string
-     * @Type("string")
-     * @SerializedName("to-charset")
-     */
-    protected $toCharset;
-
-    /**
-     * The subject's file mode for a CSV target file.
-     *
-     * @var string
-     * @Type("string")
-     * @SerializedName("file-mode")
-     */
-    protected $fileMode;
 
     /**
      * The flag to signal that the subject has to use the strict mode or not.
@@ -621,66 +586,6 @@ class Configuration implements ConfigurationInterface
     public function getMultipleValueDelimiter()
     {
         return $this->multipleValueDelimiter;
-    }
-
-    /**
-     * Return's the delimiter character to use, default value is comma (,).
-     *
-     * @return string The delimiter character
-     */
-    public function getDelimiter()
-    {
-        return $this->delimiter;
-    }
-
-    /**
-     * The enclosure character to use, default value is double quotation (").
-     *
-     * @return string The enclosure character
-     */
-    public function getEnclosure()
-    {
-        return $this->enclosure;
-    }
-
-    /**
-     * The escape character to use, default value is backslash (\).
-     *
-     * @return string The escape character
-     */
-    public function getEscape()
-    {
-        return $this->escape;
-    }
-
-    /**
-     * The file encoding of the CSV source file, default value is UTF-8.
-     *
-     * @return string The charset used by the CSV source file
-     */
-    public function getFromCharset()
-    {
-        return $this->fromCharset;
-    }
-
-    /**
-     * The file encoding of the CSV targetfile, default value is UTF-8.
-     *
-     * @return string The charset used by the CSV target file
-     */
-    public function getToCharset()
-    {
-        return $this->toCharset;
-    }
-
-    /**
-     * The file mode of the CSV target file, either one of write or append, default is write.
-     *
-     * @return string The file mode of the CSV target file
-     */
-    public function getFileMode()
-    {
-        return $this->fileMode;
     }
 
     /**